feat(api): add status blueprint
