+++ app/routes.py
@@ -1557,10 +1557,15 @@ def allowed_filename(filename: str) -> bool:
 def init_routes(app: Flask) -> None:
     """Register all route handlers on the given ``app``."""
     from app.blueprints.network import create_blueprint
+    from app.blueprints.status import create_blueprint as create_status_blueprint
 
     if not getattr(app, "_network_bp_registered", False):
         app.register_blueprint(create_blueprint())
         app._network_bp_registered = True
+
+    if not getattr(app, "_status_bp_registered", False):
+        app.register_blueprint(create_status_blueprint())
+        app._status_bp_registered = True
 
     # get_active_groups()
 
@@ -1607,227 +1612,6 @@ def init_routes(app: Flask) -> None:
             CLOCK_DIGITAL=CLOCK_DIGITAL,
             CLOCK_NAVBAR=CLOCK_NAVBAR,
         )
-
-    # Add a new route for the extended health check
-    @app.route("/health")
-    @login_required
-    @profile_route("/health")
-    def health_check():
-
-        scheduler_status = "failed"
-        free_gb = 0
-
-        metrics = scheduling.get_system_metrics()
-
-        # Define thresholds for nominal performance
-        cpu_threshold = 80  # 80% CPU usage
-        memory_threshold = 80  # 80% memory usage
-        thread_threshold = 100  # 100 threads # should be tied to the thread count in the config, right?
-        open_file_threshold = 1024  # thats a lot
-        disk_threshold = 95  # almost full
-
-        # Check if metrics are nominal and collect error messages
-        error_messages = []
-        is_nominal = True
-
-        if metrics["cpu_usage"] >= cpu_threshold:
-            is_nominal = False
-            error_messages.append(f"CPU usage is high: {metrics['cpu_usage']}%")
-        if metrics["memory_usage"] >= memory_threshold:
-            is_nominal = False
-            error_messages.append(f"Memory usage is high: {metrics['memory_usage']}%")
-        if metrics["thread_count"] >= thread_threshold:
-            is_nominal = False
-            error_messages.append(f"Thread count is high: {metrics['thread_count']}")
-        if metrics["open_files"] >= open_file_threshold:
-            is_nominal = False
-            error_messages.append(f"Too many open files: {metrics['open_files']}")
-        if metrics["disk_usage"] >= disk_threshold:
-            is_nominal = False
-            error_messages.append(f"Disk usage is high: {metrics['disk_usage']}%")
-
-        try:
-            # 0h 1m 6s
-            if (
-                len(metrics["uptime"]) < 9 and "0h 0m " in metrics["uptime"]
-            ):  # first ten seconds...
-                is_nominal = False
-                error_messages.append("System just started, still initializing")
-        except Exception:
-            is_nominal = False
-            error_messages.append("Error getting system uptime")
-
-        ######
-        #  consider rolling these into the metrics
-        try:
-            # Check database connection
-            session = SessionLocal()
-            session.execute(text("SELECT 1"))
-            session.close()
-            db_status = "connected"
-        except Exception:
-            is_nominal = False
-            db_status = "disconnected"
-            error_messages.append("Database connection failed")
-
-        try:
-            # Check if scheduler is running
-            scheduler_status = "running" if scheduling.scheduler.running else "stopped"
-            if scheduler_status != "running":
-                is_nominal = False
-                error_messages.append("Scheduler is not running")
-        except Exception:
-            is_nominal = False
-            scheduler_status = "failed"
-            error_messages.append("Error checking scheduler status")
-
-        #
-        ######
-
-        return (
-            jsonify(
-                {
-                    "status": "healthy" if is_nominal else "degraded",
-                    "metrics": metrics,
-                    "nominal": is_nominal,
-                    "database": db_status,
-                    "scheduler": scheduler_status,
-                    "free_disk_space_gb": free_gb,
-                    "error_messages": error_messages,
-                }
-            ),
-            200,
-        )  # always return 200, but might be degraded.
-
-    @app.route("/danger_status")
-    @login_required
-    @profile_route("/danger_status")
-    def danger_status():
-        """Return whether Danger mode can be used."""
-        port_open = is_chrome_debug_port_open("127.0.0.1", 9222)
-        idle = not check_user_activity(timeout=1)
-        enabled = config.get_setting("DANGER_MODE", "True") == "True"
-        browser_path = get_chrome_path()
-        patched = not shortcuts_need_patch()
-        return jsonify(
-            {
-                "port_open": port_open,
-                "idle": idle,
-                "enabled": enabled,
-                "ready": port_open and idle and enabled,
-                "browser": (os.path.basename(browser_path) if browser_path else None),
-                "path": browser_path,
-                "version": (get_chrome_version(browser_path) if browser_path else None),
-                "shortcut": str(first_shortcut_path() or ""),
-                "patched": patched,
-            }
-        )
-
-    @app.route("/captions_status")
-    @login_required
-    def captions_status():
-        """Return the most recent caption and timestamp.
-
-        If a ``group`` query parameter is provided, the newest caption from
-        templates in that group is returned. Otherwise the latest global summary
-        is used.
-        """
-
-        group = request.args.get("group")
-        if group and group != "all":
-            templates = template_manager.get_templates()
-            latest_time = None
-            caption = ""
-            for name, tmpl in templates.items():
-                groups = [g.strip() for g in tmpl.get("groups", "").split(",")]
-                if group not in groups:
-                    continue
-                t = tmpl.get("last_caption_time")
-                if not t:
-                    continue
-                try:
-                    dt = datetime.strptime(t, "%Y-%m-%d %H:%M:%S")
-                except Exception:
-                    continue
-                if not latest_time or dt > latest_time:
-                    latest_time = dt
-                    caption = tmpl.get("last_caption", "")
-            if latest_time:
-                return jsonify(
-                    {
-                        "caption": caption,
-                        "timestamp": latest_time.strftime("%Y-%m-%d %H:%M:%S"),
-                    }
-                )
-
-        caption = ""
-        timestamp = ""
-        session_db = SessionLocal()
-        try:
-            rec = session_db.query(Summary).order_by(Summary.timestamp.desc()).first()
-            if rec:
-                try:
-                    data = json.loads(getattr(rec, "content", ""))
-                    if data:
-                        caption = next(iter(data.values()))
-                except Exception:
-                    caption = getattr(rec, "content", "")
-                ts = getattr(rec, "timestamp", None)
-                if ts is not None:
-                    timestamp = datetime.utcfromtimestamp(ts).strftime(
-                        "%Y-%m-%dT%H:%M:%SZ"
-                    )
-        except Exception as e:  # pragma: no cover - unexpected DB errors
-            logging.error("error retrieving captions status: %s", e)
-        finally:
-            session_db.close()
-
-        return jsonify({"caption": caption, "timestamp": timestamp})
-
-    @app.route("/discovery_status")
-    @login_required
-    @profile_route("/discovery_status")
-    def discovery_status():
-        """Return cached background discovery status."""
-        return jsonify(scheduling.get_discovery_status())
-
-    @app.route("/discover/subnets")
-    @login_required
-    def discover_subnets():
-        """Return local IPv4 subnets as strings."""
-        nets = [str(n) for n in camera_discovery._local_subnets()]
-        nets.append("internet")
-        return jsonify(nets)
-
-    @app.route("/toggle_discovery", methods=["POST"])
-    @login_required
-    @profile_route("/toggle_discovery")
-    def toggle_discovery():
-        """Start or stop hourly background discovery."""
-        try:
-            job = scheduling.scheduler.get_job("background_discovery")
-            if job:
-                scheduling.stop_discovery()
-                update_setting("DISCOVERY_AUTOSTART", "False", restart=False)
-                return jsonify({"status": "stopped"})
-            scheduling.schedule_discovery()
-            update_setting("DISCOVERY_AUTOSTART", "True", restart=False)
-            return jsonify({"status": "running"})
-        except Exception as e:
-            return jsonify({"status": "error", "message": str(e)}), 500
-
-    @app.route("/toggle_chyron", methods=["POST"])
-    @login_required
-    @profile_route("/toggle_chyron")
-    def toggle_chyron():
-        """Enable or disable the caption chyron."""
-        try:
-            current = config.get_setting("CHYRON_SPEED", "0")
-            new_speed = "0" if str(current) != "0" else "240"
-            update_setting("CHYRON_SPEED", new_speed)
-            return jsonify({"speed": int(new_speed)})
-        except Exception as e:
-            return jsonify({"error": str(e)}), 500
 
     @app.route("/danger", methods=["GET", "POST"])
     @login_required
